requirements.txt

djangorestframework-jwt
Django==3.1.3
django-heroku==0.3.1
dj-database-url==0.5.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.0.4
whitenoise==5.2.0
python-dotenv==0.15.0
